+++ src/common/interfaces/database/EnvironmentDatabaseInterface.ts
@@ -1,7 +1,0 @@
-import EnvironmentDataInterface from '../EnvironmentDataInterface';
-import MonitorHistoryInterface from '../MonitorHistoryInterface';
-
-export default interface EnvironmentDatabaseInterface {
-  environments: EnvironmentDataInterface[];
-  monitoringHistory: MonitorHistoryInterface[];
-}